+++ lib/src/lib/components/form-fields/select-field/select-field-infinite-scroll.directive.ts
@@ -1,7 +1,8 @@
-import { AfterViewInit, Directive, EventEmitter, Input, NgZone, OnDestroy, OnInit, Output } from '@angular/core';
+import { computed, Directive, effect, inject, input, model, NgZone, output, signal } from '@angular/core';
 import { MatSelect } from '@angular/material/select';
-import { debounceTime, takeUntil, tap } from 'rxjs/operators';
-import { fromEvent, Subject } from 'rxjs';
+import { debounceTime } from 'rxjs/operators';
+import { fromEvent, Subscription } from 'rxjs';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 /** The height of the select items in `em` units. */
 const SELECT_ITEM_HEIGHT_EM = 3;
@@ -11,85 +12,70 @@ const SELECT_ITEM_HEIGHT_EM = 3;
  */
 @Directive({
   selector: 'mat-select[lab900InfiniteScroll]',
-  standalone: true,
 })
-export class SelectInfiniteScrollDirective implements OnInit, OnDestroy, AfterViewInit {
-  @Input() public threshold = '15%';
-  @Input() public debounceTime = 150;
-  @Input() public complete!: boolean;
-  @Output() public infiniteScroll = new EventEmitter<void>();
+export class SelectInfiniteScrollDirective {
+  private readonly matSelect = inject(MatSelect, { optional: false });
+  private readonly ngZone = inject(NgZone);
 
-  private panel!: Element;
-  private thrPx = 0;
-  private thrPc = 0;
-  private singleOptionHeight = SELECT_ITEM_HEIGHT_EM;
+  public readonly threshold = input<string>('15%');
+  public readonly debounceTime = input<number>(150);
+  public readonly complete = model<boolean>(false);
+  public readonly infiniteScroll = output<void>();
 
-  private destroyed$ = new Subject<boolean>();
+  private readonly panel = signal<Element | undefined>(undefined);
+  private readonly thrPx = computed(() => (this.threshold().lastIndexOf('%') > -1 ? 0 : parseFloat(this.threshold())));
+  private readonly thrPc = computed(() =>
+    this.threshold().lastIndexOf('%') > -1 ? parseFloat(this.threshold()) / 100 : 0
+  );
+  private readonly singleOptionHeight = computed(() => {
+    const panel = this.panel();
+    if (panel) {
+      return parseFloat(getComputedStyle(panel).fontSize) * SELECT_ITEM_HEIGHT_EM;
+    }
+    return SELECT_ITEM_HEIGHT_EM;
+  });
 
-  public constructor(
-    private matSelect: MatSelect,
-    private ngZone: NgZone
-  ) {}
-
-  public ngOnInit(): void {
-    this.evaluateThreshold();
-  }
-
-  public ngAfterViewInit(): void {
-    this.matSelect.openedChange.pipe(takeUntil(this.destroyed$)).subscribe(opened => {
-      if (opened) {
-        this.panel = this.matSelect.panel.nativeElement;
-        this.singleOptionHeight = this.getSelectItemHeightPx();
-        this.registerScrollListener();
+  public constructor() {
+    let sub: Subscription;
+    effect(() => {
+      const panel = this.panel();
+      if (panel) {
+        sub?.unsubscribe();
+        sub = this.registerScrollListener(panel);
       }
+      return () => {
+        sub?.unsubscribe();
+      };
+    });
+    this.matSelect.openedChange.pipe(takeUntilDestroyed()).subscribe(opened => {
+      setTimeout(() => {
+        if (opened && this.matSelect.panel) {
+          this.panel.set(this.matSelect.panel.nativeElement);
+        }
+      });
     });
   }
 
-  public ngOnDestroy(): void {
-    this.destroyed$.next(true);
-    this.destroyed$.complete();
+  private registerScrollListener(panel: Element): Subscription {
+    return fromEvent(panel, 'scroll')
+      .pipe(debounceTime(this.debounceTime()))
+      .subscribe(event => this.handleScrollEvent(panel, event));
   }
 
-  private evaluateThreshold(): void {
-    if (this.threshold.lastIndexOf('%') > -1) {
-      this.thrPx = 0;
-      this.thrPc = parseFloat(this.threshold) / 100;
-    } else {
-      this.thrPx = parseFloat(this.threshold);
-      this.thrPc = 0;
-    }
-  }
-
-  private registerScrollListener(): void {
-    fromEvent(this.panel, 'scroll')
-      .pipe(
-        takeUntil(this.destroyed$),
-        debounceTime(this.debounceTime),
-        tap(event => {
-          this.handleScrollEvent(event);
-        })
-      )
-      .subscribe();
-  }
-
-  private handleScrollEvent(event: any): void {
+  private handleScrollEvent(panel: Element, event: any): void {
     this.ngZone.runOutsideAngular(() => {
-      if (this.complete) {
+      if (this.complete()) {
         return;
       }
-      const countOfRenderedOptions = this.matSelect.options.length;
-      const infiniteScrollDistance = this.singleOptionHeight * countOfRenderedOptions;
-      const threshold = this.thrPc !== 0 ? infiniteScrollDistance * this.thrPc : this.thrPx;
+      const countOfRenderedOptions = this.matSelect.options?.length ?? 0;
+      const infiniteScrollDistance = this.singleOptionHeight() * countOfRenderedOptions;
+      const threshold = this.thrPc() !== 0 ? infiniteScrollDistance * this.thrPc() : this.thrPx();
 
-      const scrolledDistance = this.panel.clientHeight + event.target.scrollTop;
+      const scrolledDistance = panel.clientHeight + event.target.scrollTop;
 
       if (scrolledDistance + threshold >= infiniteScrollDistance) {
         this.ngZone.run(() => this.infiniteScroll.emit());
       }
     });
-  }
-
-  private getSelectItemHeightPx(): number {
-    return parseFloat(getComputedStyle(this.panel).fontSize) * SELECT_ITEM_HEIGHT_EM;
   }
 }